BFS ordenado en grafo y cola
